refactor(frontend): move BTC network utils to lib
