+++ src/frontend/src/btc/components/core/BtcLoaderWallets.svelte
@@ -1,11 +1,6 @@
 <script lang="ts">
 	import { nonNullish } from '@dfinity/utils';
 	import { initBtcWalletWorker } from '$btc/services/worker.btc-wallet.services';
-	import {
-		isNetworkIdBTCMainnet,
-		isNetworkIdBTCRegtest,
-		isNetworkIdBTCTestnet
-	} from '$icp/utils/ic-send.utils.js';
 	import LoaderWallets from '$lib/components/core/LoaderWallets.svelte';
 	import { LOCAL } from '$lib/constants/app.constants';
 	import {
@@ -16,6 +11,11 @@
 	import { enabledBtcTokens } from '$lib/derived/tokens.derived';
 	import type { InitWalletWorkerFn } from '$lib/types/listener';
 	import type { Token } from '$lib/types/token';
+	import {
+		isNetworkIdBTCMainnet,
+		isNetworkIdBTCRegtest,
+		isNetworkIdBTCTestnet
+	} from '$lib/utils/network.utils';
 
 	let tokens: Token[];
