+++ src/frontend/src/lib/utils/network.utils.ts
@@ -1,6 +1,9 @@
 import type { BitcoinNetwork as SignerBitcoinNetwork } from '$declarations/signer/signer.did';
 import {
 	BITCOIN_NETWORKS_IDS,
+	BTC_MAINNET_NETWORK_ID,
+	BTC_REGTEST_NETWORK_ID,
+	BTC_TESTNET_NETWORK_ID,
 	ICP_NETWORK_ID,
 	SUPPORTED_ETHEREUM_NETWORKS_IDS
 } from '$env/networks.env';
@@ -20,6 +23,15 @@ export const isNetworkIdEthereum = (id: NetworkId | undefined): boolean =>
 
 export const isNetworkIdBitcoin = (id: NetworkId | undefined): boolean =>
 	nonNullish(id) && BITCOIN_NETWORKS_IDS.includes(id);
+
+export const isNetworkIdBTCMainnet = (networkId: NetworkId | undefined): boolean =>
+	BTC_MAINNET_NETWORK_ID === networkId;
+
+export const isNetworkIdBTCTestnet = (networkId: NetworkId | undefined): boolean =>
+	BTC_TESTNET_NETWORK_ID === networkId;
+
+export const isNetworkIdBTCRegtest = (networkId: NetworkId | undefined): boolean =>
+	BTC_REGTEST_NETWORK_ID === networkId;
 
 /**
  * Filter the tokens that either lives on the selected network or, if no network is provided, pseud Chain Fusion, then those that are not testnets.
